Cleanup log-prelude module.
Analyzer class and Model are not supposed to be user defined,
set them statically to "Honeypot" and "Nepenthes".

The log-prelude.conf file is now optional, but might be provided
in order to configure the Nepenthes analyzer name and/or profile
(default is nepenthes).

Exit the Prelude module in case client initialization fail.

Minor fixes:
 Remove deprecated inclusion.
 Use autoconf VERSION in place of NEPENTHES_VERSION.
 Correct NEPENTHES project URL.
 Use idmef_path_new_fast().